feat(leds): rebuild the colour pipeline around light rather than duty
Every value the renderer handed the panel was a duty cycle, and duty is
not light: the three dies differ by nearly four to one, so a VAL_LOW
purple put out a third of what a VAL_LOW yellow did and a bipolar ring
read brighter positive than negative. Everything here follows from
fixing that, and from giving the framebuffer somewhere to put the
fractions it needs once it does.

The framebuffer is 8.8 fixed point and the flush dithers it with a
first-order sigma-delta, so a colour can ask for a fraction of a duty
step and get it. The dither lives in the flush rather than the
framebuffer, which is what keeps the simulator showing the ramp instead
of the noise. See led_curve.h, which is now the one file that says how a
number becomes light, split into what the device does and how a screen
should draw it.

The bipolar CV ramp drops the blue axis: red to green, nothing else. A
perceptual curve with a knee at 5V spends 70% of the visible range on
the half of the voltage most patches use, the dies are luminance matched
so equal magnitudes read equally bright either side of zero, and past 5V
the colour warms rather than changing family. It also no longer wraps a
param value past full scale into a nearly dark ring.

led_set_hsv balances every hue to the brightness it was asked for, so a
VAL_* means the same thing whatever colour wears it - the spread across
the palette went from 3.6x to flat. With the dim hues lifted, the clamp
page stopped being the only one drawn a step up and everything else was
levelled to meet it: base layers are VAL_BASE, and what has to read as
brighter than a page goes to VAL_HIG.

The simulator and web panel reconstruct display RGB from what each die
emits, dividing by the screen primaries' own luminance. Without that
step light was being ranked as if it were hue, and a blue - which really
does put out more green light than blue - was drawn as pale cyan.

The assignment marker is a travelling value-noise field sampled at each
LED's real position on the board, rather than one gate flashing
everything in unison. It washes an element's colour toward pale instead
of whitening it, and a held source's destinations wear the same field
without its rests.

The stepped-random pattern length now reads like the FRQ ring: hue for
the division's prime limit, and a pulse at the rate it produces, sharing
one scale with FRQ so a triplet is a triplet on both pages. The 7-step
length is gone - it needed a fourth colour crowded between orange and
red - which also frees 8KB of normalisation table.

+++ Core/Inc/Lib/engine_state.h
@@ -3,15 +3,20 @@
 
 #include "clock_sync.h"
 #include "hw_setup.h"
+#include "led_curve.h" // IWYU pragma: keep - LED_UNIT is the unit of the fields below
 #include <stdint.h>
 
 // The engine's running state: phases, output levels, trigger edges, the scene
 // blend, the LED framebuffer. None of it is persisted and all of it is
 // reproducible from the config plus the input stream.
 
+// 8.8 fixed point duty per primary - LED_UNIT is one step of what the WS2812
+// takes. The framebuffer holds the colour that was meant and the flush is where
+// it becomes eight bits, so the fractional part survives long enough to be
+// dithered. See led_curve.h.
 typedef struct
 {
-  uint8_t r, g, b;
+  uint16_t r, g, b;
 } LedRgb;
 
 typedef struct
